[#10] refactor: 일반 댓글과 익명 댓글 서비스 분리
