+++ src/main/java/com/mallang/comment/application/AnonymousCommentService.java
@@ -1,8 +1,6 @@
 package com.mallang.comment.application;
 
-import com.mallang.comment.application.command.WriteAuthenticatedCommentCommand;
-import com.mallang.comment.domain.AuthenticatedWriter;
-import com.mallang.comment.domain.AuthenticatedWriterRepository;
+import com.mallang.comment.application.command.WriteAnonymousCommentCommand;
 import com.mallang.comment.domain.Comment;
 import com.mallang.comment.domain.CommentRepository;
 import com.mallang.post.domain.Post;
@@ -14,16 +12,14 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional
 @Service
-public class CommentService {
+public class AnonymousCommentService {
 
     private final PostRepository postRepository;
     private final CommentRepository commentRepository;
-    private final AuthenticatedWriterRepository authenticatedWriterRepository;
 
-    public Long write(WriteAuthenticatedCommentCommand command) {
+    public Long write(WriteAnonymousCommentCommand command) {
         Post post = postRepository.getById(command.postId());
-        AuthenticatedWriter writer = authenticatedWriterRepository.getByMemberId(command.memberId());
-        Comment comment = command.toComment(post, writer);
+        Comment comment = command.toComment(post);
         Comment saved = commentRepository.save(comment);
         return saved.getId();
     }
