+++ src/main/java/com/mallang/comment/presentation/CommentController.java
@@ -1,5 +1,6 @@
 package com.mallang.comment.presentation;
 
+import com.mallang.comment.application.AnonymousCommentService;
 import com.mallang.comment.application.CommentService;
 import com.mallang.comment.presentation.request.WriteAnonymousCommentRequest;
 import com.mallang.comment.presentation.request.WriteAuthenticatedCommentRequest;
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class CommentController {
 
     private final CommentService commentService;
+    private final AnonymousCommentService anonymousCommentService;
 
     @PostMapping
     public ResponseEntity<Void> write(
@@ -32,7 +34,7 @@ public class CommentController {
     public ResponseEntity<Void> anonymousWrite(
             @RequestBody WriteAnonymousCommentRequest request
     ) {
-        Long id = commentService.anonymousWrite(request.toCommand());
+        Long id = anonymousCommentService.write(request.toCommand());
         return ResponseEntity.created(URI.create("/comments/" + id)).build();
     }
 }
